adjusted variables to project
