+++ src/constants/paths.js
@@ -1,0 +1,17 @@
+const paths = {
+	DASHBOARD: {
+		name: "Strona główna",
+		path: "/",
+	},
+	BLOG_CATEGORIES: {
+		name: "Kategorie",
+		path: "/blog/:category",
+	},
+	BLOG_ENTRY: {
+		name: "Wpis",
+		path: "/blog/:category/:entryId",
+		redirect: (category, entryId) => `/blog/${category}/${entryId}`,
+	},
+}
+
+export default paths;
