+++ src/constants/routes.js
@@ -1,18 +1,10 @@
+import BlogEntry from 'pages/BlogEntry';
+import paths from './paths';
 
-export const pathList = {
-    DASHBOARD: {
-        name: "Strona główna",
-        path: "/",
-    },
-    BLOG_CATEGORIES: {
-        name: "Kategorie",
-        path: "/blog/:category",
-    },
-    BLOG_ENTRY: {
-        name: "Wpis",
-        path: "/blog/:category/:entryId",
-        redirect: (category, entryId) => `/blog/${category}/${entryId}`,
-    },
-}
-
-export default []
+export default [
+    {
+        path: paths.BLOG_ENTRY.path,
+        component: BlogEntry,
+        exact: true
+    }
+]
