+++ src/pages/BlogEntry/index.js
@@ -1,3 +1,3 @@
-import BlogEntry from './BlogEntry';
+import BlogEntry from 'pages/BlogEntry';
 
 export default BlogEntry;
